feat(ios): MIM-231 unify Tailcat connection testing
